Удаляет вызовы состояний страницы
