+++ index.html
@@ -110,7 +110,8 @@
           <label class="ad-form__label" for="price">Цена за ночь, руб.</label>
           <input id="price" name="price" type="number" placeholder="5000" min="0" max="100000" required
           data-pristine-required-message="Обязательное поле"
-          data-pristine-minlength-message="Максимальное значение 100000">
+          data-pristine-min-message="Минимальное значение 0"
+          data-pristine-max-message="Максимальное значение 100000">
           <div class="ad-form__slider"></div>
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--time">
